% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdownwebrdemo-package.R
\docType{package}
\name{pkgdownwebrdemo-package}
\alias{pkgdownwebrdemo}
\alias{pkgdownwebrdemo-package}
\title{pkgdownwebrdemo: pkgdown webR Demo for interactive documentation}
\description{
Package to test out generating Web Assembly binaries for R packages using r-wasm GitHub Actions and deploying an interactive cell within a pkgdown documentation site.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rd.thecoatlessprofessor.com/webr-in-pkgdown}
  \item \url{https://github.com/coatless-r-n-d/webr-in-pkgdown}
  \item Report bugs at \url{https://github.com/coatless-r-n-d/webr-in-pkgdown/issues}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID})

}
\keyword{internal}
