% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in-webr.R
\name{in_webr}
\alias{in_webr}
\title{Detect if the version of R is built using Web Assembly (WASM)}
\usage{
in_webr()
}
\description{
The function checks whether the compiled version of \emph{R} shows an architecture
that matches \code{"emscripten"} . If it's the case, we're likely using webR to access
R.
}
\section{WebR}{
\ifelse{html}{\out{
<p><a href="https://webr.r-wasm.org/latest/#code=W3sibmFtZSI6ImV4YW1wbGUuUiIsInBhdGgiOiIvZXhhbXBsZS5SIiwidGV4dCI6IiMgQ2hlY2sgdG8gc2VlIGlmIFdBU00gaXMgYWN0aXZlXG5pbl93ZWJyKCkifV0\%3D&ju" target="_blank" style="background-color: #007bff; color: white; padding: 8px 16px; text-decoration: none; border-radius: 4px; font-size: 14px;">🌐 View in webR REPL</a></p>
}}{\ifelse{latex}{\url{https://webr.r-wasm.org/latest/#code=W3sibmFtZSI6ImV4YW1wbGUuUiIsInBhdGgiOiIvZXhhbXBsZS5SIiwidGV4dCI6IiMgQ2hlY2sgdG8gc2VlIGlmIFdBU00gaXMgYWN0aXZlXG5pbl93ZWJyKCkifV0\%3D&ju}}{Interactive webR content not available for this output format.}}
}

\examples{
# Check to see if WASM is active
in_webr()
}
