% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_surrealism.R
\docType{data}
\name{residual_surrealism}
\alias{residual_surrealism}
\title{Residual (Sur)Realism}
\format{
A data frame with 5,395 rows and 7 columns:
\describe{
\item{V1}{Target variable of interest}
\item{V2}{First predictor variable}
\item{V3}{Second predictor variable}
\item{V4}{Third predictor variable}
\item{V5}{Fourth predictor variable}
\item{V6}{Fiveth predictor variable}
\item{V7}{Sixth predictor variable}
}
}
\source{
Leonard A Stefanski
\url{https://www.tandfonline.com/doi/abs/10.1198/000313007X190079}
}
\usage{
residual_surrealism
}
\description{
A dataset that yields a model with surreal residuals.
}
\section{WebR}{
\ifelse{html}{\out{
<p><a href="https://webr.r-wasm.org/latest/#code=W3sibmFtZSI6ImV4YW1wbGUuUiIsInBhdGgiOiIvZXhhbXBsZS5SIiwidGV4dCI6IiMgTG9hZCB0aGUgZGF0YVxuZGF0YShyZXNpZHVhbF9zdXJyZWFsaXNtKVxuXG4jIEZpdCBhIG1vZGVsXG5tb2RlbCA9IGxtKFYxIH4gLiwgcmVzaWR1YWxfc3VycmVhbGlzbSlcblxuIyBHcmFwaCB0aGUgcmVzaWR1YWxzIHZzLiBmaXR0ZWQgdmFsdWVzXG5wbG90KGZpdHRlZChtb2RlbCkscmVzaWQobW9kZWwpKVxuXG4jIEFkZCBhIGxpbmUgYXQgeSA9IDBcbmFibGluZSgwLCAwLCBjb2wgPSBcInJlZFwiKSJ9XQ\%3D\%3D&ju" target="_blank" style="background-color: #007bff; color: white; padding: 8px 16px; text-decoration: none; border-radius: 4px; font-size: 14px;">🌐 View in webR REPL</a></p>
}}{\ifelse{latex}{\url{https://webr.r-wasm.org/latest/#code=W3sibmFtZSI6ImV4YW1wbGUuUiIsInBhdGgiOiIvZXhhbXBsZS5SIiwidGV4dCI6IiMgTG9hZCB0aGUgZGF0YVxuZGF0YShyZXNpZHVhbF9zdXJyZWFsaXNtKVxuXG4jIEZpdCBhIG1vZGVsXG5tb2RlbCA9IGxtKFYxIH4gLiwgcmVzaWR1YWxfc3VycmVhbGlzbSlcblxuIyBHcmFwaCB0aGUgcmVzaWR1YWxzIHZzLiBmaXR0ZWQgdmFsdWVzXG5wbG90KGZpdHRlZChtb2RlbCkscmVzaWQobW9kZWwpKVxuXG4jIEFkZCBhIGxpbmUgYXQgeSA9IDBcbmFibGluZSgwLCAwLCBjb2wgPSBcInJlZFwiKSJ9XQ\%3D\%3D&ju}}{Interactive webR content not available for this output format.}}
}

\examples{
# Load the data
data(residual_surrealism)

# Fit a model
model = lm(V1 ~ ., residual_surrealism)

# Graph the residuals vs. fitted values
plot(fitted(model),resid(model))

# Add a line at y = 0
abline(0, 0, col = "red")
}
\keyword{datasets}
